+++ src/bcf/annotate_dgidb.rs
@@ -137,7 +137,6 @@ fn modify_vcf_entries(
             }
         }
     }
-
     Ok(())
 }